refactor: move Massive provider to provider/massive sub-package
