+++ cmd/providers_register.go
@@ -18,4 +18,5 @@ package cmd
 import (
 	_ "github.com/penny-vault/pvdata/provider/fred"
 	_ "github.com/penny-vault/pvdata/provider/legacy"
+	_ "github.com/penny-vault/pvdata/provider/massive"
 )
